compose: Move call link insertion to new compose_call_ui module.
