+++ tools/test-js-with-node
@@ -64,6 +64,7 @@ EXEMPT_FILES = make_set(
         "web/src/compose.js",
         "web/src/compose_actions.js",
         "web/src/compose_banner.ts",
+        "web/src/compose_call_ui.js",
         "web/src/compose_closed_ui.js",
         "web/src/compose_fade.js",
         "web/src/compose_popovers.js",
